+++ Core/include/CFOpenDirectory/CFODContext.h
@@ -1,0 +1,42 @@
+/*
+ * Copyright (c) 2009 Apple Inc. All rights reserved.
+ *
+ * @APPLE_LICENSE_HEADER_START@
+ * 
+ * This file contains Original Code and/or Modifications of Original Code
+ * as defined in and that are subject to the Apple Public Source License
+ * Version 2.0 (the 'License'). You may not use this file except in
+ * compliance with the License. Please obtain a copy of the License at
+ * http://www.opensource.apple.com/apsl/ and read it before using this
+ * file.
+ * 
+ * The Original Code and all software distributed under the License are
+ * distributed on an 'AS IS' basis, WITHOUT WARRANTY OF ANY KIND, EITHER
+ * EXPRESS OR IMPLIED, AND APPLE HEREBY DISCLAIMS ALL SUCH WARRANTIES,
+ * INCLUDING WITHOUT LIMITATION, ANY WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE, QUIET ENJOYMENT OR NON-INFRINGEMENT.
+ * Please see the License for the specific language governing rights and
+ * limitations under the License.
+ * 
+ * @APPLE_LICENSE_HEADER_END@
+ */
+
+#if !defined(__OPENDIRECTORY_CFODCONTEXT__)
+#define __OPENDIRECTORY_CFODCONTEXT__ 1
+
+#include <CFOpenDirectory/CFOpenDirectory.h>
+
+__BEGIN_DECLS
+
+/*!
+    @function   ODContextGetTypeID
+    @abstract   Standard GetTypeID function support for CF-based objects
+    @discussion Returns the typeID for the ODContext object
+    @result     a valid CFTypeID for the ODContext object
+*/
+CF_EXPORT
+CFTypeID ODContextGetTypeID(void);
+
+__END_DECLS
+
+#endif /* ! __OPENDIRECTORY_CFODCONTEXT__ */
