+++ Core/include/CFOpenDirectory/record_internal.h
@@ -1,0 +1,42 @@
+/*
+ * Copyright (c) 2010 Apple Inc. All rights reserved.
+ *
+ * @APPLE_LICENSE_HEADER_START@
+ * 
+ * This file contains Original Code and/or Modifications of Original Code
+ * as defined in and that are subject to the Apple Public Source License
+ * Version 2.0 (the 'License'). You may not use this file except in
+ * compliance with the License. Please obtain a copy of the License at
+ * http://www.opensource.apple.com/apsl/ and read it before using this
+ * file.
+ * 
+ * The Original Code and all software distributed under the License are
+ * distributed on an 'AS IS' basis, WITHOUT WARRANTY OF ANY KIND, EITHER
+ * EXPRESS OR IMPLIED, AND APPLE HEREBY DISCLAIMS ALL SUCH WARRANTIES,
+ * INCLUDING WITHOUT LIMITATION, ANY WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE, QUIET ENJOYMENT OR NON-INFRINGEMENT.
+ * Please see the License for the specific language governing rights and
+ * limitations under the License.
+ * 
+ * @APPLE_LICENSE_HEADER_END@
+ */
+
+#include <CoreFoundation/CFBridgingPriv.h>
+
+struct __ODRecord {
+	CFRuntimeBase _base;
+	ODSessionRef _session;
+	ODNodeRef _orignode;
+	ODNodeRef _realnode;
+
+	CFSetRef _attrset;
+	CFMutableDictionaryRef _attributes;
+	dispatch_queue_t _attrq;
+
+	/* purely for convenience */
+	CFStringRef _type;
+	CFStringRef _name;
+	CFStringRef _metaname;
+};
+
+ODNodeRef _ODRecordGetNode(ODRecordRef record);
